u: move http tests into http test module
