+++ python-web-server/tests/http/requests/deferred/valid/compat_obs_fold_huge.py
@@ -1,4 +1,4 @@
-from tests.treq import uri
+from tests.http.treq import uri
 from web_server.config import MessageConfig
 
 cfg = MessageConfig.default()
